added producer files and sample
