+++ AnaProd/tasks.py
@@ -78,7 +78,7 @@ class InputFileTask(Task, law.LocalWorkflow):
 
 
 class AnaTupleTask(Task, HTCondorWorkflow, law.LocalWorkflow):
-    max_runtime = copy_param(HTCondorWorkflow.max_runtime, 20.0)
+    max_runtime = copy_param(HTCondorWorkflow.max_runtime, 5.0)
 
     def workflow_requires(self):
         return { "anaCache" : AnaCacheTask.req(self, branches=()), "inputFile": InputFileTask.req(self,workflow='local', branches=()) }
@@ -92,7 +92,7 @@ class AnaTupleTask(Task, HTCondorWorkflow, law.LocalWorkflow):
         n = 0
         branches = {}
         for sample_id, sample_name in enumerate(sorted(self.samples.keys())):
-            inputFileTxt = InputFileTask.req(self, branch=sample_id,workflow='local', branches=()).output().path
+            inputFileTxt = InputFileTask.req(self, branch=sample_id,workflow='local', branches=(sample_id,)).output().path
             with open(inputFileTxt, 'r') as inputtxtFile:
                 input_files = inputtxtFile.read().splitlines()
             if len(input_files) == 0:
@@ -100,11 +100,6 @@ class AnaTupleTask(Task, HTCondorWorkflow, law.LocalWorkflow):
             for input_file in input_files:
                 branches[n] = (sample_id, sample_name, self.samples[sample_name]['sampleType'], input_file)
                 n += 1
-                '''
-                if n >= len(self.branches):
-                    print(branches)
-                    return branches
-                '''
         return branches
 
     def output(self, force_pre_output=False):
@@ -133,8 +128,8 @@ class AnaTupleTask(Task, HTCondorWorkflow, law.LocalWorkflow):
         if sample_type!='data':
             skimtuple_cmd = ['python3', producer_skimtuples, '--inputDir',outdir_anatuples, '--centralFile',outFileName, '--workingDir', outdir_skimtuples,
                      '--outputFile', outFileName]
-            if self.test:
-                skimtuple_cmd.extend(['--test' , 'True'])
+            #if self.test:
+                #skimtuple_cmd.extend(['--test' , 'True'])
             sh_call(skimtuple_cmd,verbose=1)
         tmpFile = os.path.join(outdir_skimtuples, outFileName)
         if sample_type=='data':
@@ -160,50 +155,92 @@ class AnaTupleTask(Task, HTCondorWorkflow, law.LocalWorkflow):
         return os.path.join(central_anaTuples_path, sample_name)
 
 class DataMergeTask(Task, HTCondorWorkflow, law.LocalWorkflow):
-    max_runtime = copy_param(HTCondorWorkflow.max_runtime, 25.0)
-    max_files_per_job = luigi.IntParameter(default=1000000, description="maximum number of input files per job")
+    max_runtime = copy_param(HTCondorWorkflow.max_runtime, 5.0)
+
+    def workflow_requires(self):
+        prod_branches = self.create_branch_map()
+        workflow_dict = {}
+        workflow_dict["anaTuple"] = {
+            idx: AnaTupleTask.req(self, branches=tuple((br,) for br in branches))
+            for idx, branches in prod_branches.items()
+        }
+        print(prod_branches)
+        return workflow_dict
 
     def requires(self):
-        prod_branches, input_files = self.branch_data
-        print(self.branch_data)
+        prod_branches = self.branch_data
         deps = [AnaTupleTask.req(self, max_runtime=AnaCacheTask.max_runtime._default, branch=prod_br) for prod_br in prod_branches ]
         return deps
 
 
     def create_branch_map(self):
-        self.load_sample_configs()
         deps = []
         anaProd_branch_map = AnaTupleTask.req(self, branch=-1, branches=()).create_branch_map()
         prod_branches = []
-        input_files = []
         for prod_br, (sample_id, sample_name, sample_type, input_file) in anaProd_branch_map.items():
-            if sample_type == "data":
-                print(input_file)
-                if input_file not in input_files:
-                    input_files.append(input_file)
+            if sample_type != "data": continue
             prod_branches.append(prod_br)
-        return { 0: (prod_branches, input_files) }
-
-
+        return { 0: prod_branches }
 
     def output(self, force_pre_output=False):
         out = os.path.join(self.central_anaTuples_path(), 'data','nano.root')
         return law.LocalFileTarget(out)
 
     def run(self):
-        prod_branches, input_files = self.branch_data
-        print(input_files)
-        job_home, remove_job_home = self.law_job_home()
-        input_files_str = ' '.join(input_files)
-        outdir_dataMerge = os.path.join(job_home, 'data')
+        print("starting run")
+        print(self.input)
+        '''
+        prod_branches = self.branch_data
+        outdir_dataMerge = os.path.join(self.central_anaTuples_path(), 'data')
         producer_dataMerge = os.path.join(self.ana_path(), 'AnaProd', 'MergeNtuples.py')
         finalFile = self.output().path
         tmpFile = os.path.join(outdir_dataMerge, 'data.root')
-        dataMerge_cmd = ['python3', producer_dataMerge, input_files_str, '--outFile', tmpFile ]
+        dataMerge_cmd = ['python3', producer_dataMerge, self.input, '--outFile', tmpFile ]
         sh_call(dataMerge_cmd,verbose=1)
         if self.test: print(f"finalFile is {finalFile}")
         shutil.copy(tmpFile, finalFile)
         if os.path.exists(finalFile):
             os.remove(tmpFile)
-        if remove_job_home:
-            shutil.rmtree(job_home)
+        '''
+
+class HistProducerTask(Task, HTCondorWorkflow, law.LocalWorkflow):
+    max_runtime = copy_param(HTCondorWorkflow.max_runtime, 5.0)
+
+    def workflow_requires(self):
+        return { "anaTuple" : AnaTupleTask.req(self, branches=())}
+
+    def requires(self):
+        sample_name, sample_id = self.branch_data
+        return [ AnaTupleTask.req(self, branch=sample_id, max_runtime=AnaTupleTask.max_runtime._default, branches=())]
+
+    def create_branch_map(self):
+        self.load_sample_configs()
+        n = 0
+        branches = {}
+        anaProd_branch_map = AnaTupleTask.req(self, branch=-1, branches=()).create_branch_map()
+        for prod_br, (sample_id, sample_name, sample_type, input_file) in anaProd_branch_map.items():
+            if sample_type =='data': continue
+            branches[n] =  (sample_name, sample_id)
+            n+=1
+        return branches
+
+    def output(self):
+        sample_name, sample_id = self.branch_data
+        if not os.path.isdir(self.central_Histograms_path()):
+            os.makedirs(self.central_Histograms_path())
+        out = os.path.join(self.central_Histograms_path(), sample_name)
+        return law.LocalDirectoryTarget(out)
+
+    def run(self):
+        sample_name, sample_id = self.branch_data
+        print(sample_name)
+        '''
+        inDir = os.path.join(self.central_anaTuples_path(), sample_name)
+        HistProducer = os.path.join(self.ana_path(), 'Analysis', 'HistProducer.py')
+        outDir = self.output().path
+        if os.path.isdir(finalDir):
+            os.remove(finalDir)
+        print(self.input)
+        HistProducer_cmd = ['python3', HistProducer,'--inputDir', inDir, '--dataset', sample_name, '--outDir', finalDir , '--compute_unc_variations', 'True', '--compute_rel_weights', 'True']
+        sh_call(HistProducer_cmd,verbose=1)
+        '''
