+++ run_tools/law_customizations.py
@@ -62,6 +62,28 @@ def select_items(all_items, filters):
     return list(sorted(selected_items))
 
 
+global_params = None
+samples = None
+
+def load_sample_configs(sample_config, period):
+    global global_params
+    global samples
+
+    if global_params is None:
+        with open(sample_config, 'r') as f:
+            samples = yaml.safe_load(f)
+
+        global_params = samples['GLOBAL']
+        all_samples = []
+        for key, value in samples.items():
+            if(type(value) != dict):
+                raise RuntimeError(f'Invalid sample definition period="{period}", sample_name="{key}"' )
+            if key != 'GLOBAL':
+                all_samples.append(key)
+        selected_samples = select_items(all_samples, global_params.get('sampleSelection', []))
+        samples = { key : samples[key] for key in selected_samples }
+    return global_params, samples
+
 
 class Task(law.Task):
     """
@@ -78,6 +100,8 @@ class Task(law.Task):
         super(Task, self).__init__(*args, **kwargs)
         self.cmssw_env_ = None
         self.sample_config = os.path.join(self.ana_path(), 'config', f'samples_{self.period}.yaml')
+        self.global_params, self.samples = load_sample_configs(self.sample_config, self.period)
+
 
     def load_sample_configs(self):
         with open(self.sample_config, 'r') as f:
@@ -113,6 +137,9 @@ class Task(law.Task):
 
     def central_anaTuples_path(self):
         return os.path.join(self.central_path(), 'anaTuples', self.period, self.version)
+
+    def central_Histograms_path(self):
+        return os.path.join(self.central_path(), 'histograms', self.period, self.version)
 
     def central_anaCache_path(self):
         return os.path.join(self.central_path(), 'anaCache', self.period)
